# Optimize findMaxLandMassSize() which can take 10% of AI time.