--- a/game-core/src/main/java/games/strategy/triplea/ai/pro/util/ProTerritoryValueUtils.java
+++ b/game-core/src/main/java/games/strategy/triplea/ai/pro/util/ProTerritoryValueUtils.java
@@ -469,10 +469,7 @@ public final class ProTerritoryValueUtils {
 
               @Override
               public boolean shouldContinueSearch(final int distanceSearched) {
-                if (distanceSearched >= MIN_FACTORY_CHECK_DISTANCE && !found.isEmpty()) {
-                  return false;
-                }
-                return true;
+                return distanceSearched < MIN_FACTORY_CHECK_DISTANCE || found.isEmpty();
               }
             });
     return found;