Optimize findMaxLandMassSize() which can take 10% of AI time.
    
This change optimizes findMaxLandMassSize() by reducing the amount of work needed to do. The main difference is we now keep track of land territories already processed so that we don't need to start new searches from them.
    
This creates a new BreadthFirstSearch class that takes out the BFS implementation I implemented earlier for finding capitals, so that it can be re-used here.
    
It also fixes a correctness issue with finding the max land size - the previous search would only search up to a distance of 6. The new implementation does not have this limitation.
    
On Domination 1914 map, which has a lot of territories, this saves about 50 seconds of CPU time over a round of an all-AI game.

## Functional Changes
<!-- Put an X next any that apply -->
[] New map or map update
[] New Feature
[] Feature update or enhancement
[] Feature Removal
[] Code Cleanup or refactor
[] Configuration Change
[] Problem fix:  <!-- Link to bug issue or forum post here -->
[X] Other: Optimization

## Testing
Tested manually on a game of Domination 1914 and profiled the game.