@RoiEXLab 2020-04-11
Overall I like the changes @asvitkine 👍 
Performance improvements by faster algorithms are always a good thing