@DanVanAtta 2020-04-11
> Optimize findMaxLandMassSize() which can take 10% of AI time.

> On Domination 1914 map, which has a lot of territories, this saves about 50 seconds of CPU time over a round of an all-AI game.

@asvitkine does that imply you observed an AI time reduction from 500s to 450s?