@DanVanAtta 2020-04-12
@asvitkine before pushing changes to this branch, please see if the CI not building this PR is potentially resolved by: https://github.com/triplea-game/triplea/pull/6200#issuecomment-612643835