@asvitkine 2020-04-13
> > Optimize findMaxLandMassSize() which can take 10% of AI time.
> 
> > On Domination 1914 map, which has a lot of territories, this saves about 50 seconds of CPU time over a round of an all-AI game.
> 
> @asvitkine does that imply you observed an AI time reduction from 500s to 450s?

Roughly, yes.

I just re-did the measurement with all three of my optimizations applied and I'm seeing a reduction of total round time for that saved game of 110 seconds (~560s -> ~440s) and the time spent by each battle calc worker from ~120 seconds to ~90 seconds.