@DanVanAtta 2020-04-14
@asvitkine 
```
> Task :game-core:checkstyleMain

[ant:checkstyle] [WARN] /home/travis/build/triplea-game/triplea/game-core/src/main/java/games/strategy/engine/data/util/BreadthFirstSearch.java:26:32: Redundant 'final' modifier. [RedundantModifier]
```