@asvitkine 2020-04-16
Done, sorry about that!

On Wed, Apr 15, 2020 at 11:22 PM Dan Van Atta <notifications@github.com>
wrote:

> *@DanVanAtta* commented on this pull request.
> ------------------------------
>
> In
> game-core/src/main/java/games/strategy/engine/data/util/BreadthFirstSearch.java
> <https://github.com/triplea-game/triplea/pull/6187#discussion_r409257983>:
>
> > +import java.util.ArrayDeque;
> +import java.util.HashSet;
> +import java.util.List;
> +import java.util.Set;
> +import java.util.function.Predicate;
> +import org.triplea.java.ObjectUtils;
> +
> +/**
> + * Implements Breadth First Search (BFS) to traverse / find territories. Since the search criteria
> + * varies depending on the use case, the class is designed to be sub-classed, with methods visit()
> + * and shouldContinueSearch() that can be overridden to customize the behavior.
> + */
> +public class BreadthFirstSearch {
> +  public abstract static class Visitor {
> +    /**
> +     * Called when a new territory is encountered. Can be overridden to provide custom search
>
> @asvitkine <https://github.com/asvitkine> I'm not seeing the latest
> updates, do they need to be pushed perhaps?
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/triplea-game/triplea/pull/6187#discussion_r409257983>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAEJ4GKXAEYHCABE5CMAVLRMZ2YRANCNFSM4MGAHN3Q>
> .
>