@DanVanAtta 2020-04-16
@asvitkine 

```
> Task :game-core:checkstyleTest

[ant:checkstyle] [WARN] /home/travis/build/triplea-game/triplea/game-core/build/resources/test/minimap.xml:526:1: Line contains a tab character. [FileTabCharacter]

[ant:checkstyle] [WARN] /home/travis/build/triplea-game/triplea/game-core/build/resources/test/minimap.xml:527:1: Line contains a tab character. [FileTabCharacter]

> Task :game-core:checkstyleTest FAILED

> Task :game-headed-javafx:compileJava

FAILURE: Build failed with an exception.

* What went wrong:

Execution failed for task ':game-core:checkstyleTest'.

> Checkstyle rule violations were found. See the report at: file:///home/travis/build/triplea-game/triplea/game-core/build/reports/checkstyle/test.html

  Checkstyle files with violations: 1

  Checkstyle violations by severity: [warning:2]
```