@asvitkine 2020-04-16
Thanks, fixed.

On Thu, Apr 16, 2020 at 4:08 AM Dan Van Atta <notifications@github.com>
wrote:

> @asvitkine <https://github.com/asvitkine>
>
> > Task :game-core:checkstyleTest
>
> [ant:checkstyle] [WARN] /home/travis/build/triplea-game/triplea/game-core/build/resources/test/minimap.xml:526:1: Line contains a tab character. [FileTabCharacter]
>
> [ant:checkstyle] [WARN] /home/travis/build/triplea-game/triplea/game-core/build/resources/test/minimap.xml:527:1: Line contains a tab character. [FileTabCharacter]
>
> > Task :game-core:checkstyleTest FAILED
>
> > Task :game-headed-javafx:compileJava
>
> FAILURE: Build failed with an exception.
>
> * What went wrong:
>
> Execution failed for task ':game-core:checkstyleTest'.
>
> > Checkstyle rule violations were found. See the report at: file:///home/travis/build/triplea-game/triplea/game-core/build/reports/checkstyle/test.html
>
>   Checkstyle files with violations: 1
>
>   Checkstyle violations by severity: [warning:2]
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/triplea-game/triplea/pull/6187#issuecomment-614485893>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAAEJ4FZ7DLZSGCKTOL5LYTRM24GXANCNFSM4MGAHN3Q>
> .
>