@DanVanAtta 2020-04-16
side note:

> * Try to reduce complexity to appease codeclimate

One should not appease codeclimate, the commit comment is actually not very useful. What is done to reduce complexity? What is the change made? Appeasing codeclimate is a means to an end. For example "Fix checkstyles" is not useful either, but "improveme javadoc, fix completness of javadoc, fix whitespace formatting" are more useful.